# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing users, authentication, and clinic operations.",
        "title": "Dental Clinic API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/ai/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processes a user chat message, updates booking state, and creates an appointment when state is complete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Assistant"
                ],
                "summary": "AI assistant chat",
                "parameters": [
                    {
                        "description": "Chat message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/schedule/doctors/{doctorId}/working-hours": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a doctor schedule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get doctor schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID (UUID)",
                        "name": "doctorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ScheduleDoctorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Create new schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID (UUID)",
                        "name": "doctorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateScheduleResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new slots",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Generate new slots",
                "parameters": [
                    {
                        "description": "Generate slots data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateSlotsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/working-hours/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Update doctor schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete doctor schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ChatRequest": {
            "type": "object",
            "properties": {
                "choice_id": {
                    "type": "string"
                },
                "choice_type": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ChatResponse": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "string"
                },
                "available_slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_ai_assistant_dto.SlotResponse"
                    }
                },
                "choice_required": {
                    "type": "boolean"
                },
                "choice_type": {
                    "type": "string"
                },
                "clinics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ClinicOption"
                    }
                },
                "doctors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DoctorOption"
                    }
                },
                "reply": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServiceOption"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/models.BookingState"
                }
            }
        },
...
        "dto.CreateScheduleRequest": {
            "type": "object",
            "properties": {
                "clinic_address_id": {
                    "description": "Doctor_id   \t\t\tstring    `json:\"doctor_id\"`",
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "dto.CreateScheduleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "success": {
                    "type": "string"
                }
            }
        },
...
        "dto.GenerateSlotsRequest": {
            "type": "object",
            "properties": {
                "from_date": {
                    "type": "string"
                },
                "to_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.ScheduleDoctorResponse": {
            "type": "object",
            "properties": {
                "clinic_address_id": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateScheduleRequest": {
            "type": "object",
            "properties": {
                "clinic_address_id": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "doctor_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "internal_modules_ai_assistant_dto.SlotResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "slot_end": {
                    "type": "string"
                },
                "slot_start": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.BookingState": {
            "type": "object",
            "properties": {
                "clinic_address_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "step": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.ClinicOption": {
            "type": "object",
            "properties": {
                "clinic_address_id": {
                    "type": "string"
                },
                "clinic_id": {
                    "type": "string"
                },
                "clinic_name": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.DoctorOption": {
            "type": "object",
            "properties": {
                "experience": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
        "models.ServiceOption": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}